# Speed up `read_fgd_dem()` by using data.table