--- a/NEWS.md
+++ b/NEWS.md
@@ -1,5 +1,6 @@
 # fgdr (development version)
 
+* Speed up `read_fgd_dem()` by using data.table for backend.
 * Added support for terra format output with `read_fgd_dem()` 
 * When a DEM file is read and returned as a df, a unit (meter) attribute is added to the elevation value.