--- a/R/commons.R
+++ b/R/commons.R
@@ -79,26 +79,25 @@ fgd_file_info <- function(file, ...) {
 }
 
 fgd_dem_file_info <- function(file, ...) {
-
   file_info <-
     fgd_file_info(file, ...)
-
   is5m <-
     file_info$xml_docs %>%
     xml2::xml_find_first("/*/*[3]/*[6]") %>% # nolint
     xml2::xml_contents() %>%
     as.character() %>%
     stringr::str_detect("5m")
-
   meshcode <-
     file_info$xml_docs %>%
     xml2::xml_find_all("/*/*[3]/*[7]") %>% # nolint
     xml2::xml_contents() %>%
     as.character()
-
+  line_startdend <-
+    which(grepl("gml:tupleList", readLines(file)))
   list(xml_docs = file_info$xml_docs,
        type = file_info$type,
        is5m = is5m,
+       skip_n = line_startdend[1],
+       limit_n = line_startdend[2] - line_startdend[1] - 1,
        meshcode = meshcode)
-
 }