--- a/tests/testthat/test-read_fgd_dem.R
+++ b/tests/testthat/test-read_fgd_dem.R
@@ -1,17 +1,26 @@
 context("test-read_fgd_dem")
 
 test_that("Successed on dummies", {
-  res <-
+  res_df <-
     read_fgd_dem(file = system.file("extdata/FG-GML-0000-00-00-DEM5A-dummy.xml",
                                     package = "fgdr"),
                  resolution = 5,
-                 return_class = "df")
-  expect_is(res, "data.frame")
-  expect_identical(as.character(res[1, 1]),
+                 return_class = "data.frame")
+  expect_is(res_df, "data.frame")
+  expect_identical(as.character(res_df[1, 1]),
                    paste(intToUtf8(c(12381, 12398, 20182), multiple = TRUE), collapse = ""))
-  expect_s3_class(res, "tbl_df")
-  expect_equal(dim(res), c(33750, 2))
-  expect_named(res, c("type", "value"))
+  expect_s3_class(res_df, "tbl_df")
+  expect_equal(dim(res_df), c(33750, 2))
+  expect_named(res_df, c("type", "value"))
+  res_dt <-
+    read_fgd_dem(file = system.file("extdata/FG-GML-0000-00-00-DEM5A-dummy.xml",
+                                    package = "fgdr"),
+                 resolution = 5,
+                 return_class = "data.table")
+  expect_is(res_dt, "data.table")
+  expect_equal(dim(res_dt), dim(res_df))
+  expect_equal(res_dt$type, res_df$type)
+  expect_equal(res_dt$value, res_df$value)
 
   res <-
     read_fgd_dem(system.file("extdata/FG-GML-0000-00-00-DEM5A-dummy.xml",
@@ -25,7 +34,7 @@ test_that("Successed on dummies", {
     read_fgd_dem(system.file("extdata/FG-GML-0000-10-dem10b-dummy.xml",
                              package = "fgdr"),
                  resolution = 10,
-                 return_class = "df")
+                 return_class = "data.frame")
   expect_is(res, "data.frame")
   expect_s3_class(res, "tbl_df")
   expect_equal(dim(res), c(843750, 2))