## Summary

- data.tableでの読み込みによる高速化
- data.frameでの出力もオプションとして残す
    - `return_class`の指定を `data.frame`, `data.table`のように分ける

### Benchmark

```r
library(fgdr)
library(bench)

res_current_5m <-
  mark(
    current = read_fgd_dem(
      "data-raw/FG-GML-5135-63-DEM5A/FG-GML-5135-63-00-DEM5A-20161001.xml",
      resolution = 5,
      return_class = "df"
    )
  )
res_current_10m <-
  mark(
    current = read_fgd_dem(
      "data-raw/FG-GML-5440-10-dem10b-20161001.xml",
      resolution = 10,
      return_class = "df"
    )
  )


pkgload::load_all()
res_dt_5m <-
  mark(
    data_table = read_fgd_dem(
      "data-raw/FG-GML-5135-63-DEM5A/FG-GML-5135-63-00-DEM5A-20161001.xml",
      resolution = 5,
      return_class = "data.table"
    )
  )
res_dt_10m <-
  mark(
    data_table = read_fgd_dem(
      "data-raw/FG-GML-5440-10-dem10b-20161001.xml",
      resolution = 10,
      return_class = "data.table"
    )
  )


rbind(
  res_current_5m,
  res_dt_5m)
#> # A tibble: 2 x 13
#> expression     min  median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result          memory         time     gc       
#> <bch:expr> <bch:t> <bch:t>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list>          <list>         <list>   <list>   
#> 1 current     32.9ms  34.4ms      28.9    10.2MB     4.45    13     2      450ms <tibble [33,75… <Rprofmem[,3]… <bch:tm… <tibble …
#> 2 data_table  28.5ms  29.8ms      27.3    6.77MB     4.20    13     2      476ms <data.table[,2… <Rprofmem[,3]… <bch:tm… <tibble …

rbind(
  res_current_10m,
  res_dt_10m)
#> # A tibble: 2 x 13
#> expression     min  median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result           memory         time    gc       
#> <bch:expr> <bch:t> <bch:t>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list>           <list>         <list>  <list>   
#> 1 current      967ms   967ms      1.03     136MB     3.10     1     3      967ms <tibble [843,75… <Rprofmem[,3]… <bch:t… <tibble …
#> 2 data_table   836ms   836ms      1.20      92MB     4.78     1     4      836ms <data.table[,2]… <Rprofmem[,3]… <bch:t… <tibble …
```

## Related issues

#16 のリベンジ